Change report generation tests to check task not management command
